+++ hikvision-doorbell/src/doorbell.py
@@ -485,8 +485,8 @@ class Doorbell():
                 pass
 
         # We have run out of available endpoints to call, dont ro a runtime error, just continue with 0 outputs
-        logger.debug("Unable to get the number of doors on the indoor station, please configure the relays manually with this option in the config: output_relays")
-        return 0
+        logger.debug("Unable to get the number of doors on the indoor station, please configure the relays manually with this option in the config: output_relays, we will continue with 1 output relay as a fallback")
+        return 1
         #raise RuntimeError("Unable to get the number of doors, please configure the relays manually with this option in the config: output_relays")
 
     def get_num_outputs(self) -> int:
@@ -573,8 +573,8 @@ class Doorbell():
                 # This endpoint failed, try the next one
                 pass
         # We have run out of available endpoints to call, dont ro a runtime error, just continue with 0 outputs
-        logger.info("Unable to get the number of doors, please configure the relays manually with this option in the config: output_relays")
-        return 0
+        logger.info("Unable to get the number of doors, please configure the relays manually with this option in the config: output_relays, we will continue with 1 output relay as a fallback")
+        return 1
         #raise RuntimeError("Unable to get the number of doors, please configure the relays manually with this option in the config: output_relays")
 
     def get_num_coms_indoor(self) -> int: